[Bug #157712291] fix code coverage
