+++ .babelrc
@@ -1,5 +1,10 @@
 {
   "presets": [
       "env"
-  ]
+  ],
+  "env": {
+    "test": {
+      "plugins": [ "istanbul" ]
+    }
+  }
 }
