+++ package.json
@@ -7,10 +7,26 @@
     "start": "babel-node server",
     "dev": "concurrently \"npm run start:dev\" \"npm run lint:watch\"",
     "start:dev": "cross-env NODE_ENV=development nodemon --exec babel-node -- server",
-    "test": "mocha --require babel-core/register server/test/**/*.js --exit",
-    "test:watch": "cross-env NODE_ENV=test nodemon --exec \"mocha --recursive --require babel-core/register server/test/**/*.js\"",
+    "test": "cross-env NODE_ENV=test nyc mocha server/test/**/*.js",
+    "test:watch": "npm test -- --watch",
     "lint": "./node_modules/.bin/esw  server",
     "lint:watch": "npm run lint -- --watch"
+  },
+  "nyc": {
+    "require": [
+      "babel-register"
+    ],
+    "reporter": [
+      "lcov",
+      "text",
+      "html"
+    ],
+    "sourceMap": false,
+    "instrument": false,
+    "exclude": [
+      "server/test",
+      "server/db"
+    ]
   },
   "repository": {
     "type": "git",
@@ -33,6 +49,7 @@
   "devDependencies": {
     "babel-cli": "^6.26.0",
     "babel-core": "^6.26.3",
+    "babel-plugin-istanbul": "^4.1.6",
     "babel-preset-env": "^1.7.0",
     "babel-register": "^6.26.0",
     "concurrently": "^3.5.1",
@@ -46,6 +63,7 @@
     "mocha": "^5.1.1",
     "mocha-lcov-reporter": "^1.3.0",
     "node-mocks-http": "^1.5.8",
-    "nodemon": "^1.17.4"
+    "nodemon": "^1.17.4",
+    "nyc": "^11.8.0"
   }
 }
