+++ server/test/routes/requests.test.js
@@ -1,6 +1,6 @@
 import chai from 'chai';
 import chaiHttp from 'chai-http';
-import server from '../../index';
+import server from '../../app';
 import requests from '../../db/requests';
 import testData from '../testData';
